+++ R/visualization.R
@@ -264,26 +264,37 @@ DoHeatmap <- function(
   # group.use <- factor(x = group.use[cells])
   plots <- vector(mode = 'list', length = ncol(x = groups.use))
   for (i in 1:ncol(x = groups.use)) {
+    data.group <- data
     group.use <- groups.use[, i, drop = TRUE]
-    group.use <- factor(x = group.use)
+    if (!is.factor(x = group.use)) {
+      group.use <- factor(x = group.use)
+    }
     names(x = group.use) <- cells
     if (draw.lines) {
       # create fake cells to serve as the white lines, fill with NAs
-      lines.width <- lines.width %||% ceiling(x = nrow(x = data) * 0.0025)
+      lines.width <- lines.width %||% ceiling(x = nrow(x = data.group) * 0.0025)
       placeholder.cells <- sapply(
         X = 1:(length(x = levels(x = group.use)) * lines.width),
-        FUN = function(x) RandomName(length = 20)
+        FUN = function(x) {
+          return(RandomName(length = 20))
+        }
       )
       placeholder.groups <- rep(x = levels(x = group.use), times = lines.width)
+      group.levels <- levels(x = group.use)
       names(x = placeholder.groups) <- placeholder.cells
       group.use <- as.vector(x = group.use)
       names(x = group.use) <- cells
-      group.use <- factor(x = c(group.use, placeholder.groups))
-      na.data <- matrix(data = NA, nrow = length(x = placeholder.cells), ncol = ncol(data), dimnames = list(placeholder.cells, colnames(x = data)))
-      data <- rbind(data, na.data)
+      group.use <- factor(x = c(group.use, placeholder.groups), levels = group.levels)
+      na.data.group <- matrix(
+        data = NA,
+        nrow = length(x = placeholder.cells),
+        ncol = ncol(x = data.group),
+        dimnames = list(placeholder.cells, colnames(x = data.group))
+      )
+      data.group <- rbind(data.group, na.data.group)
     }
     plot <- SingleRasterMap(
-      data = data,
+      data = data.group,
       raster = raster,
       disp.min = disp.min,
       disp.max = disp.max,
@@ -563,7 +574,9 @@ VlnPlot <- function(
 #' @param object Seurat object
 #' @param dims Dimensions to plot, must be a two-length numeric vector specifying x- and y-dimensions
 #' @param cells Vector of cells to plot (default is all cells)
-#' @param cols Vector of colors, each color corresponds to an identity class. By default, ggplot2 assigns colors
+#' @param cols Vector of colors, each color corresponds to an identity class. This may also be a single character
+#' or numeric value corresponding to a palette as specified by \code{\link[RColorBrewer]{brewer.pal.info}}.
+#' By default, ggplot2 assigns colors
 #' @param pt.size Adjust point size for plotting
 #' @param reduction Which dimensionality reduction to use. If not specified, first searches for umap, then tsne, then pca
 #' @param group.by Name of one or more metadata columns to group (color) cells by
@@ -626,7 +639,7 @@ DimPlot <- function(
   label.size = 4,
   repel = FALSE,
   cells.highlight = NULL,
-  cols.highlight = 'red',
+  cols.highlight = '#DE2D26',
   sizes.highlight = 1,
   na.value = 'grey50',
   combine = TRUE,
@@ -636,12 +649,7 @@ DimPlot <- function(
   if (length(x = dims) != 2) {
     stop("'dims' must be a two-length vector")
   }
-  reduction <- reduction %||% {
-    default.reductions <- c('umap', 'tsne', 'pca')
-    object.reductions <- FilterObjects(object = object, classes.keep = 'DimReduc')
-    reduc.use <- min(which(x = default.reductions %in% object.reductions))
-    default.reductions[reduc.use]
-  }
+  reduction <- reduction %||% DefaultDimReduc(object = object)
   cells <- cells %||% colnames(x = object)
   data <- Embeddings(object = object[[reduction]])[cells, dims]
   data <- as.data.frame(x = data)
@@ -690,10 +698,10 @@ DimPlot <- function(
         plot <- plot + FacetTheme() +
           facet_wrap(
             facets = vars(!!sym(x = split.by)),
-            ncol = if (length(x = group.by) > 1) {
+            ncol = if (length(x = group.by) > 1 || is.null(x = ncol)) {
               length(x = unique(x = data[, split.by]))
             } else {
-              NULL
+              ncol
             }
           )
       }
@@ -742,10 +750,12 @@ DimPlot <- function(
 #' @param ncol Number of columns to combine multiple feature plots to, ignored if \code{split.by} is not \code{NULL}
 #' @param combine Combine plots into a single gg object; note that if TRUE; themeing will not work when plotting multiple features
 #' @param coord.fixed Plot cartesian coordinates with fixed aspect ratio
-#' @param by.col If splitting by a factor, plot the splits per column with the features as rows.
+#' @param by.col If splitting by a factor, plot the splits per column with the features as rows; ignored if \code{blend = TRUE}
 
 #'
-#' @return A ggplot object
+#' @return Returns a ggplot object if only 1 feature is plotted.
+#' If >1 features are plotted and \code{combine=TRUE}, returns a combined ggplot object using \code{cowplot::plot_grid}.
+#' If >1 features are plotted and \code{combine=FALSE}, returns a list of ggplot objects.
 #'
 #' @importFrom grDevices rgb
 #' @importFrom cowplot theme_cowplot
@@ -771,7 +781,7 @@ FeaturePlot <- function(
   features,
   dims = c(1, 2),
   cells = NULL,
-  cols = c("lightgrey",  "blue"),
+  cols =  ifelse(test = c(blend, blend), yes = c("#ff0000", "#00ff00"), no = c('lightgrey', 'blue')),
   pt.size = NULL,
   order = FALSE,
   min.cutoff = NA,
@@ -784,6 +794,7 @@ FeaturePlot <- function(
   blend.threshold = 0.5,
   label = FALSE,
   label.size = 4,
+  repel = FALSE,
   ncol = NULL,
   combine = TRUE,
   coord.fixed = FALSE,
@@ -810,10 +821,34 @@ FeaturePlot <- function(
   if (blend && length(x = features) != 2) {
     stop("Blending feature plots only works with two features")
   }
+  if (blend) {
+    if (length(x = cols) > 2) {
+      warning(
+        "Blending feature plots only works with two colors; using first two colors",
+        call. = FALSE,
+        immediate. = TRUE
+      )
+    } else if (length(x = cols) < 2) {
+      warning(
+        "Blended feature plots require two colors, using default colors",
+        call. = FALSE,
+        immediate. = TRUE
+      )
+      cols <- c("#ff0000", "#00ff00")
+    }
+  }
+  if (blend && length(x = cols) != 2) {
+    stop("Blending feature plots only works with two colors")
+  }
   dims <- paste0(Key(object = object[[reduction]]), dims)
   cells <- cells %||% colnames(x = object)
-  data <- FetchData(object = object, vars = c(dims, features), cells = cells, slot = slot)
-  if (ncol(x = data) < 3) {
+  data <- FetchData(
+    object = object,
+    vars = c(dims, 'ident', features),
+    cells = cells,
+    slot = slot
+  )
+  if (ncol(x = data) < 4) {
     stop(
       "None of the requested features were found: ",
       paste(features, collapse = ', '),
@@ -824,7 +859,7 @@ FeaturePlot <- function(
   } else if (!all(dims %in% colnames(x = data))) {
     stop("The dimensions requested were not found", call. = FALSE)
   }
-  features <- colnames(x = data)[3:ncol(x = data)]
+  features <- colnames(x = data)[4:ncol(x = data)]
   min.cutoff <- mapply(
     FUN = function(cutoff, feature) {
       return(ifelse(
@@ -860,12 +895,12 @@ FeaturePlot <- function(
     yes = brewer.pal.info[cols, ]$maxcolors,
     no = length(x = cols)
   )
-  data[, 3:ncol(x = data)] <- sapply(
-    X = 3:ncol(x = data),
+  data[, 4:ncol(x = data)] <- sapply(
+    X = 4:ncol(x = data),
     FUN = function(index) {
       data.feature <- as.vector(x = data[, index])
-      min.use <- SetQuantile(cutoff = min.cutoff[index - 2], data.feature)
-      max.use <- SetQuantile(cutoff = max.cutoff[index - 2], data.feature)
+      min.use <- SetQuantile(cutoff = min.cutoff[index - 3], data.feature)
+      max.use <- SetQuantile(cutoff = max.cutoff[index - 3], data.feature)
       data.feature[data.feature < min.use] <- min.use
       data.feature[data.feature > max.use] <- max.use
       if (brewer.gran == 2) {
@@ -883,7 +918,7 @@ FeaturePlot <- function(
       return(data.cut)
     }
   )
-  colnames(x = data)[3:ncol(x = data)] <- features
+  colnames(x = data)[4:ncol(x = data)] <- features
   rownames(x = data) <- cells
   data$split <- if (is.null(x = split.by)) {
     RandomName()
@@ -912,7 +947,10 @@ FeaturePlot <- function(
   ylims <- c(floor(min(data[, dims[2]])), ceiling(x = max(data[, dims[2]])))
   if (blend) {
     ncol <- 4
-    color.matrix <- BlendMatrix(col.threshold = blend.threshold)
+    color.matrix <- BlendMatrix(
+      two.colors = cols,
+      col.threshold = blend.threshold
+    )
     colors <- list(
       color.matrix[, 1],
       color.matrix[1, ],
@@ -923,8 +961,16 @@ FeaturePlot <- function(
     ident <- levels(x = data$split)[i]
     data.plot <- data[as.character(x = data$split) == ident, , drop = FALSE]
     if (blend) {
-      data.plot <- cbind(data.plot[, dims], BlendExpression(data = data.plot[, features[1:2]]))
-      features <- colnames(x = data.plot)[3:ncol(x = data.plot)]
+      no.expression <- features[colMeans(x = data.plot[, features]) == 0]
+      if (length(x = no.expression) != 0) {
+        stop(
+          "The following features have no value: ",
+          paste(no.expression, collapse = ', '),
+          call. = FALSE
+        )
+      }
+      data.plot <- cbind(data.plot[, c(dims, 'ident')], BlendExpression(data = data.plot[, features[1:2]]))
+      features <- colnames(x = data.plot)[4:ncol(x = data.plot)]
     }
     for (j in 1:length(x = features)) {
       feature <- features[j]
@@ -935,19 +981,26 @@ FeaturePlot <- function(
         cols.use <- NULL
       }
       plot <- SingleDimPlot(
-        data = data.plot[, c(dims, feature, shape.by)],
+        data = data.plot[, c(dims, 'ident', feature, shape.by)],
         dims = dims,
         col.by = feature,
         order = order,
         pt.size = pt.size,
         cols = cols.use,
         shape.by = shape.by,
-        label = label,
-        label.size = label.size
+        label = FALSE
       ) +
         scale_x_continuous(limits = xlims) +
         scale_y_continuous(limits = ylims) +
         theme_cowplot()
+      if (label) {
+        plot <- LabelClusters(
+          plot = plot,
+          id = 'ident',
+          repel = repel,
+          size = label.size
+        )
+      }
       if (length(x = levels(x = data$split)) > 1) {
         plot <- plot + theme(panel.border = element_rect(fill = NA, colour = 'black'))
         plot <- plot + if (i == 1) {
@@ -983,12 +1036,21 @@ FeaturePlot <- function(
       }
       if (!blend) {
         plot <- plot + guides(color = NULL)
+        cols.grad <- cols
         if (length(x = cols) == 1) {
           plot <- plot + scale_color_brewer(palette = cols)
         } else if (length(x = cols) > 1) {
+          if (all(data.plot[, feature] == data.plot[, feature][1])) {
+            warning("All cells have the same value (", data.plot[1, feature], ") of ", feature, ".")
+            if (data.plot[1, feature][1] == 0) {
+              cols.grad <- cols[1]
+            } else{
+            cols.grad <- cols
+            }
+          }
           plot <- suppressMessages(
             expr = plot + scale_color_gradientn(
-              colors = cols,
+              colors = cols.grad,
               guide = "colorbar"
             )
           )
@@ -1055,13 +1117,19 @@ FeaturePlot <- function(
     } else {
       split.by %iff% 'none'
     }
-    if (by.col & !is.null(x = split.by)) {
-      plots <- lapply(X = plots, FUN = function(x) {
-        suppressMessages(x +
-        theme_cowplot() + ggtitle("") +
-        scale_y_continuous(sec.axis = dup_axis(name = "")) + no.right
-        )
-      })
+    if (by.col && !is.null(x = split.by) && !blend) {
+      plots <- lapply(
+        X = plots,
+        FUN = function(x) {
+          return(suppressMessages(
+            expr = x +
+              theme_cowplot() +
+              ggtitle("") +
+              scale_y_continuous(sec.axis = dup_axis(name = "")) +
+              no.right
+          ))
+        }
+      )
       nsplits <- length(x = levels(x = data$split))
       idx <- 1
       for (i in (length(x = features) * (nsplits - 1) + 1):(length(x = features) * nsplits)) {
@@ -1089,8 +1157,7 @@ FeaturePlot <- function(
         ncol = nsplits,
         legend = legend
       )
-    }
-    else {
+    } else {
       plots <- CombinePlots(
         plots = plots,
         ncol = ncol,
@@ -1671,6 +1738,9 @@ DotPlot <- function(
     Idents(object = object)
   } else {
     object[[group.by, drop = TRUE]]
+  }
+  if (!is.factor(x = data.features$id)) {
+    data.features$id <- factor(x = data.features$id)
   }
   id.levels <- levels(x = data.features$id)
   data.features$id <- as.vector(x = data.features$id)
@@ -3033,28 +3103,50 @@ BlendMap <- function(color.matrix) {
 #
 # @param n Dimensions of blended matrix (n x n)
 # @param col.threshold The color cutoff from weak signal to strong signal; ranges from 0 to 1.
+# @param two.colors Two colors used for the blend expression.
 #
 # @return An n x n matrix of blended colors
 #
-#' @importFrom grDevices rgb
+#' @importFrom grDevices rgb colorRamp
 #
-BlendMatrix <- function(n = 10, col.threshold = 0.5) {
+BlendMatrix <- function(
+  n = 10,
+  col.threshold = 0.5,
+  two.colors=c("#ff0000", "#00ff00")
+) {
   if (0 > col.threshold || col.threshold > 1) {
     stop("col.threshold must be between 0 and 1")
   }
-  return(outer(
-    X = 1:n,
-    Y = 1:n,
-    FUN = function(i, j) {
-      red <- 1 / (1 + exp(x = -(i - col.threshold * n) / 0.9))
-      green <- 1 / (1 + exp(x = -(j - col.threshold * n) / 0.9))
-      blue <- 0.2
-      alpha <- lapply(X = list(red, green, blue), FUN = '^', 40)
-      alpha <- Reduce(f = '+', x = alpha)
-      alpha <- 0.99 - 0.1 * exp(x = -alpha / 1)
-      return(rgb(red = red, green = green, blue = blue, alpha = alpha))
+  ramp <- colorRamp(colors = two.colors)
+  C1 <- ramp(x = 0)
+  C2 <- ramp(x = 1)
+  merge.weight <- min(255 / (C1 + C2 + 0.01))
+  weight_color <- function(w1, c1) {
+    c1_weight <- 1 / (1 + exp(x = -(w1 ^ 1.2 - 3 - col.threshold * 10)))
+    return(c1_weight * c1)
+  }
+  blend_color <- function(i, j) {
+    C1_weight <- weight_color(w1 = i, c1 = C1)
+    C2_weight <- weight_color(w1 = j, c1 = C2)
+    C_blend <- (merge.weight * C1_weight + merge.weight * C2_weight)
+    alpha <- lapply(X = list(i, j), FUN = '^', 0.5)
+    alpha <- Reduce(f = '+', x = alpha)
+    alpha <- (1 - 0.4 /alpha ) * 255
+    return(rgb(
+      red = C_blend[, 1],
+      green = C_blend[, 2],
+      blue = C_blend[, 3],
+      alpha = alpha,
+      maxColorValue = 255
+    ))
+  }
+  blend_matrix <- matrix(nrow = n, ncol = n)
+  for (i in 1:n) {
+    for (j in 1:n) {
+      blend_matrix[i, j] <- blend_color(i = i, j = j)
     }
-  ))
+  }
+  return(blend_matrix)
 }
 
 # Convert R colors to hexadecimal
@@ -3087,6 +3179,59 @@ Col2Hex <- function(...) {
     alpha = alpha
   )
   return(colors)
+}
+
+# Find the default DimReduc
+#
+# Searches for DimReducs matching 'umap', 'tsne', or 'pca', case-insensitive, and
+# in that order. Priority given to DimReducs matching the DefaultAssay or assay specified
+# (eg. 'pca' for the default assay weights higher than 'umap' for a non-default assay)
+#
+# @param object A Seurat object
+# @param assay Name of assay to use; defaults to the default assay of the object
+#
+# @return The default DimReduc, if possible
+#
+DefaultDimReduc <- function(object, assay = NULL) {
+  assay <- assay %||% DefaultAssay(object = object)
+  drs.use <- c('umap', 'tsne', 'pca')
+  dim.reducs <- FilterObjects(object = object, classes.keep = 'DimReduc')
+  drs.assay <- Filter(
+    f = function(x) {
+      return(DefaultAssay(object = object[[x]]) == assay)
+    },
+    x = dim.reducs
+  )
+  if (length(x = drs.assay) > 0) {
+    index <- lapply(
+      X = drs.use,
+      FUN = grep,
+      x = drs.assay,
+      ignore.case = TRUE
+    )
+    index <- Filter(f = length, x = index)
+    if (length(x = index) > 0) {
+      return(drs.assay[min(index[[1]])])
+    }
+  }
+  index <- lapply(
+    X = drs.use,
+    FUN = grep,
+    x = dim.reducs,
+    ignore.case = TRUE
+  )
+  index <- Filter(f = length, x = index)
+  if (length(x = index) < 1) {
+    stop(
+      "Unable to find a DimReduc matching one of '",
+      paste(drs.use[1:(length(x = drs.use) - 1)], collapse = "', '"),
+      "', or '",
+      drs.use[length(x = drs.use)],
+      "', please specify a dimensional reduction to use",
+      call. = FALSE
+    )
+  }
+  return(dim.reducs[min(index[[1]])])
 }
 
 # Plot feature expression by identity
@@ -3578,8 +3723,8 @@ PointLocator <- function(plot, recolor = TRUE, dark.theme = FALSE, ...) {
   points.located <- points.all[which(x = points.all$pip == 1), ]
   #   If we're recoloring, do the recolor
   if (recolor) {
-    no <- ifelse(test = dark.theme, yes = 'white', no = 'black')
-    points.all$color <- ifelse(test = points.all$pip == 1, yes = 'red', no = no)
+    no <- ifelse(test = dark.theme, yes = 'white', no = '#C3C3C3')
+    points.all$color <- ifelse(test = points.all$pip == 1, yes = '#DE2D26', no = no)
     plot.data$color <- points.all$color
     PlotBuild(data = plot.data, dark.theme = dark.theme, ...)
   }
@@ -3865,8 +4010,9 @@ SingleCorPlot <- function(
 # @param dims A two-length numeric vector with dimensions to use
 # @param pt.size Adjust point size for plotting
 # @param col.by ...
-# @param cols Vector of colors, each color corresponds to an identity class. By default, ggplot
-# assigns colors.
+# @param cols Vector of colors, each color corresponds to an identity class. This may also be a single character
+# or numeric value corresponding to a palette as specified by \code{\link[RColorBrewer]{brewer.pal.info}}.
+# By default, ggplot2 assigns colors
 # @param shape.by If NULL, all points are circles (default). You can specify any cell attribute
 # (that can be pulled with FetchData) allowing for both different colors and different shapes on
 # cells.
@@ -3888,6 +4034,7 @@ SingleCorPlot <- function(
 # @param na.value Color value for NA points when using custom scale.
 #
 #' @importFrom cowplot theme_cowplot
+#' @importFrom RColorBrewer brewer.pal.info
 #' @importFrom ggplot2 ggplot aes_string labs geom_text guides
 #' scale_color_brewer scale_color_manual element_rect guide_legend
 #'
@@ -3903,7 +4050,7 @@ SingleDimPlot <- function(
   repel = FALSE,
   label.size = 4,
   cells.highlight = NULL,
-  cols.highlight = 'red',
+  cols.highlight = '#DE2D26',
   sizes.highlight = 1,
   na.value = 'grey50'
 ) {
@@ -3925,7 +4072,7 @@ SingleDimPlot <- function(
       cells.all = rownames(x = data),
       sizes.highlight = sizes.highlight %||% pt.size,
       cols.highlight = cols.highlight,
-      col.base = cols[1] %||% 'black',
+      col.base = cols[1] %||% '#C3C3C3',
       pt.size = pt.size
     )
     order <- highlight.info$plot.order
@@ -3935,7 +4082,7 @@ SingleDimPlot <- function(
     cols <- highlight.info$color
   }
   if (!is.null(x = order) && !is.null(x = col.by)) {
-    if (typeof(x = order) == "logical"){
+    if (typeof(x = order) == "logical") {
       if (order) {
         data <- data[order(data[, col.by]), ]
       }
@@ -3991,7 +4138,7 @@ SingleDimPlot <- function(
     )
   }
   if (!is.null(x = cols)) {
-    plot <- plot + if (length(x = cols) == 1) {
+    plot <- plot + if (length(x = cols) == 1 && (is.numeric(x = cols) || cols %in% rownames(x = brewer.pal.info))) {
       scale_color_brewer(palette = cols, na.value = na.value)
     } else {
       scale_color_manual(values = cols, na.value = na.value)